修复因ydui更新而导致的数据显示问题；修复搜索栏空白时数据不读取数据的问题；进油计划列表滚动刷新；decimal改double以显示3位小数；

+++ MarineFuellingStation/Models/Store.cs
@@ -29,7 +29,7 @@ namespace MFS.Models
         /// <summary>
         /// 最近测量密度
         /// </summary>
-        public decimal Density { get; set; }
+        public double Density { get; set; }
         /// <summary>
         /// 最近测量时间
         /// </summary>
